+++ regy/src/features/tables/TableItem.js
@@ -1,15 +1,16 @@
 import React from 'react';
 import './tableItem.css';
-import { selectTable, clearSelectedTable } from '../selection/selectionSlice';
-import { removeTable } from './tablesSlice';
+import { selectTable, clearSelectedTable, selectSelectedTableId  } from '../selection/selectionSlice';
+import { removeTable, selectTables } from './tablesSlice';
 import { removeRowsOfTable } from '../rows/rowsSlice';
 import { useDispatch, useSelector } from 'react-redux';
 import { useState, useEffect } from 'react';
-import { selectSelectedTableId } from '../selection/selectionSlice';
+
 
 function TableItem({ table }) {
   const [selected, setSelected] = useState(false);
   const selectedTableId = useSelector(selectSelectedTableId);
+  const tables = useSelector(selectTables);
   const dispatch = useDispatch();
   const tableSelectHandler = (e) => {
     e.preventDefault();
@@ -21,10 +22,14 @@ function TableItem({ table }) {
     e.preventDefault();
     e.stopPropagation();
     const id = table.id;
-    dispatch(clearSelectedTable());
-    dispatch(removeRowsOfTable({tableId: id}));
-    setSelected(false);
-    dispatch(removeTable({id}))
+    const tableTitle = tables[id].title;
+    const confirmDelete = window.confirm(`Do You Really want To delete the ${tableTitle} Table?`);
+    if (confirmDelete) {
+      dispatch(clearSelectedTable());
+      dispatch(removeRowsOfTable({tableId: id}));
+      setSelected(false);
+      dispatch(removeTable({id}))
+    }
   };
 
   useEffect(() => {
